Move TupleOperations to api.tuple package
