+++ src/main/java/io/tarantool/driver/api/tuple/operations/TupleOperationBitwiseXor.java
@@ -1,31 +1,31 @@
-package io.tarantool.driver.protocol.operations;
+package io.tarantool.driver.api.tuple.operations;
 
 
 /**
- * Represents bitwise AND operation on tuple field
+ * Represents bitwise XOR operation on tuple field
  *
  * @author Sergey Volgin
  */
-public class TupleOperationBitwiseAnd extends TupleUpdateOperation {
+public class TupleOperationBitwiseXor extends TupleUpdateOperation {
 
-    public TupleOperationBitwiseAnd(int fieldIndex, long value) {
-        super(TarantoolUpdateOperationType.BITWISEAND, fieldIndex, value);
+    public TupleOperationBitwiseXor(int fieldIndex, long value) {
+        super(TarantoolUpdateOperationType.BITWISEXOR, fieldIndex, value);
         checkValue(value);
     }
 
-    public TupleOperationBitwiseAnd(String fieldName, long value) {
-        super(TarantoolUpdateOperationType.BITWISEAND, fieldName, value);
+    public TupleOperationBitwiseXor(String fieldName, long value) {
+        super(TarantoolUpdateOperationType.BITWISEXOR, fieldName, value);
         checkValue(value);
     }
 
-    private TupleOperationBitwiseAnd(TarantoolUpdateOperationType operationType, Integer fieldIndex,
+    private TupleOperationBitwiseXor(TarantoolUpdateOperationType operationType, Integer fieldIndex,
                                      String fieldName, Object value, boolean isProxyOperation) {
         super(operationType, fieldIndex, fieldName, value, isProxyOperation);
     }
 
     @Override
     public TupleOperation toProxyTupleOperation() {
-        return new TupleOperationBitwiseAnd(
+        return new TupleOperationBitwiseXor(
                 this.getOperationType(),
                 this.getFieldNumber(),
                 this.getFieldName(),
