+++ src/main/java/io/tarantool/driver/api/tuple/operations/TupleOperationSet.java
@@ -1,4 +1,4 @@
-package io.tarantool.driver.protocol.operations;
+package io.tarantool.driver.api.tuple.operations;
 
 
 /**
