+++ src/main/java/io/tarantool/driver/api/tuple/operations/package-info.java
@@ -3,4 +3,4 @@
  *
  * @author Sergey Volgin
  */
-package io.tarantool.driver.protocol.operations;
+package io.tarantool.driver.api.tuple.operations;
