+++ src/main/java/io/tarantool/driver/protocol/requests/TarantoolUpsertRequest.java
@@ -8,7 +8,7 @@ import io.tarantool.driver.protocol.TarantoolRequest;
 import io.tarantool.driver.protocol.TarantoolRequestBody;
 import io.tarantool.driver.protocol.TarantoolRequestFieldType;
 import io.tarantool.driver.protocol.TarantoolRequestType;
-import io.tarantool.driver.protocol.operations.TupleOperations;
+import io.tarantool.driver.api.tuple.operations.TupleOperations;
 
 import java.util.HashMap;
 import java.util.List;
